+++ python/GafferSceneTest/ParentTest.py
@@ -35,6 +35,7 @@
 ##########################################################################
 
 import os.path
+import inspect
 
 import imath
 
@@ -651,6 +652,109 @@ class ParentTest( GafferSceneTest.SceneTestCase ) :
 		self.assertEqual( set( str( x ) for x in parent["out"].setNames() ), { "balls", "squareThings" } )
 		self.assertEqual( parent["out"].set( "balls" ).value, IECore.PathMatcher( [ "/a/sphere" ] ) )
 		self.assertEqual( parent["out"].set( "squareThings" ).value, IECore.PathMatcher( [ "/b/box" ] ) )
+
+	def testDestination( self ) :
+
+		# /group
+		#    /sphere1
+		#	 /sphere2
+
+		script = Gaffer.ScriptNode()
+
+		script["sphere1"] = GafferScene.Sphere()
+		script["sphere1"]["name"].setValue( "sphere1" )
+		script["sphere1"]["transform"]["translate"]["x"].setValue( 1 )
+
+		script["sphere2"] = GafferScene.Sphere()
+		script["sphere2"]["name"].setValue( "sphere2" )
+		script["sphere2"]["transform"]["translate"]["x"].setValue( 2 )
+
+		script["group"] = GafferScene.Group()
+		script["group"]["in"][0].setInput( script["sphere1"]["out"] )
+		script["group"]["in"][1].setInput( script["sphere2"]["out"] )
+
+		# "Parenting" a cube to each sphere, but putting the results at
+		# the root of the scene. Using an expression to vary the dimensions
+		# and sets of each cube.
+
+		script["spheresFilter"] = GafferScene.PathFilter()
+		script["spheresFilter"]["paths"].setValue( IECore.StringVectorData( [ "/group/sphere*" ] ) )
+
+		script["cube"] = GafferScene.Cube()
+
+		script["expression"] = Gaffer.Expression()
+		script["expression"].setExpression( inspect.cleandoc(
+			"""
+			first = context.get( "parent", "" ) == "/group/sphere1"
+			parent["cube"]["sets"] = "set1" if first else "set2"
+			parent["cube"]["dimensions"]["x"] = 1 if first else 2
+			"""
+		) )
+
+		script["parent"] = GafferScene.Parent()
+		script["parent"]["in"].setInput( script["group"]["out"] )
+		script["parent"]["children"][0].setInput( script["cube"]["out"] )
+		script["parent"]["filter"].setInput( script["spheresFilter"]["out"] )
+		script["parent"]["parentVariable"].setValue( "parent" )
+		script["parent"]["destination"].setValue( "${scene:path}/../.." )
+
+		self.assertSceneValid( script["parent"]["out"] )
+
+		# Because two cubes are being added below one location, the second will
+		# have a numeric suffix to keep it unique from the other. It's ambiguous
+		# as to which one should be the second, so we define it by sorting them
+		# based on their original parent.
+
+		self.assertEqual( script["parent"]["out"].childNames( "/" ), IECore.InternedStringVectorData( [ "group", "cube", "cube1" ] ) )
+		self.assertEqual( script["parent"]["out"].object( "/cube" ).bound().size().x, 1 )
+		self.assertEqual( script["parent"]["out"].object( "/cube1" ).bound().size().x, 2 )
+		self.assertEqual( script["parent"]["out"].childNames( "/group" ), IECore.InternedStringVectorData( [ "sphere1", "sphere2" ] ) )
+		self.assertEqual( script["parent"]["out"].childNames( "/cube" ), IECore.InternedStringVectorData() )
+		self.assertEqual( script["parent"]["out"].childNames( "/cube1" ), IECore.InternedStringVectorData() )
+
+		# The contents of the sets should reflect the same sorting and uniquefying.
+
+		self.assertEqual( script["parent"]["out"].setNames(), IECore.InternedStringVectorData( [ "set1", "set2" ] ) )
+
+		self.assertEqual(
+			script["parent"]["out"].set( "set1" ).value,
+			IECore.PathMatcher( [ "/cube" ] )
+		)
+		self.assertEqual(
+			script["parent"]["out"].set( "set2" ).value,
+			IECore.PathMatcher( [ "/cube1" ] )
+		)
+
+		# We want the cubes to be positioned as if they were parented below the spheres.
+
+		self.assertEqual( script["parent"]["out"].fullTransform( "/cube" ), script["parent"]["in"].fullTransform( "/group/sphere1" ) )
+		self.assertEqual( script["parent"]["out"].fullTransform( "/cube1" ), script["parent"]["in"].fullTransform( "/group/sphere2" ) )
+
+		# And if we move the cubes to a different location, we want all that to apply still.
+
+		script["parent"]["destination"].setValue( "/group" )
+
+		self.assertSceneValid( script["parent"]["out"] )
+
+		self.assertEqual( script["parent"]["out"].childNames( "/" ), IECore.InternedStringVectorData( [ "group" ] ) )
+		self.assertEqual( script["parent"]["out"].childNames( "/group" ), IECore.InternedStringVectorData( [ "sphere1", "sphere2", "cube", "cube1" ] ) )
+		self.assertEqual( script["parent"]["out"].object( "/group/cube" ).bound().size().x, 1 )
+		self.assertEqual( script["parent"]["out"].object( "/group/cube1" ).bound().size().x, 2 )
+		self.assertEqual( script["parent"]["out"].childNames( "/group/cube" ), IECore.InternedStringVectorData() )
+		self.assertEqual( script["parent"]["out"].childNames( "/group/cube1" ), IECore.InternedStringVectorData() )
+
+		self.assertEqual( script["parent"]["out"].setNames(), IECore.InternedStringVectorData( [ "set1", "set2" ] ) )
+		self.assertEqual(
+			script["parent"]["out"].set( "set1" ).value,
+			IECore.PathMatcher( [ "/group/cube" ] )
+		)
+		self.assertEqual(
+			script["parent"]["out"].set( "set2" ).value,
+			IECore.PathMatcher( [ "/group/cube1" ] )
+		)
+
+		self.assertEqual( script["parent"]["out"].fullTransform( "/group/cube" ), script["parent"]["in"].fullTransform( "/group/sphere1" ) )
+		self.assertEqual( script["parent"]["out"].fullTransform( "/group/cube1" ), script["parent"]["in"].fullTransform( "/group/sphere2" ) )
 
 if __name__ == "__main__":
 	unittest.main()