+++ python/GafferSceneUI/ParentUI.py
@@ -95,6 +95,24 @@ Gaffer.Metadata.registerNode(
 
 		],
 
+		"destination" : [
+
+			"description",
+			"""
+			The location where the children will be placed in the output scene.
+			The default is to place the children under the parent, but they may
+			be relocated anywhere while still inheriting the parent's transform.
+			This is particularly useful when parenting lights to geometry but
+			wanting to group them and control their visibility separately.
+
+			When the destination is evaluated, the `${scene:path}` variable holds
+			the source location matched by the filter. This allows the children
+			to be placed relative to the "parent". For example, `${scene:path}/..`
+			will place the children alongside the "parent" rather than under it.
+			""",
+
+		],
+
 	}
 
 )